Flere forenklinger i style, opprydding
