+++ .gitignore
@@ -4,3 +4,4 @@
 .envrc
 ent-tavla-dev-875a70280651.json
 ent-tavla-prd-54ef424ea2f0.json
+**/.idea
