+++ tavla/app/page.tsx
@@ -7,7 +7,7 @@ import Image from 'next/image'
 import type { BoardDB } from 'src/types/db-types/boards'
 import { getBoardLinkServer } from 'src/utils/boardLink'
 import { getUserFromSessionCookie } from './(admin)/utils/server'
-import { FeatureShowcase } from './components/FeatureShowcase/FeatureShowcase'
+import { FeatureShowcase } from './components/FeatureShowcase'
 import { ImageCarousel } from './components/ImageCarousel/ImageCarousel'
 import { NavigateToOversiktButton } from './components/NavigateToOversiktButton'
 import { PreviewCarousel } from './components/PreviewCarousel'
